fix: Windows build compatibility (v12.0.1)
- Add NOMINMAX to prevent Windows min/max macro conflicts with std::min/std::max
- Add static_cast<size_t> for int64_t to size_t conversions to fix MSVC warnings

+++ ixwebsocket/IXHttpClient.cpp
@@ -487,7 +487,7 @@ namespace ix
             ss << contentLengthIt->second;
             ss >> contentLength;
 
-            auto chunkResult = _socket->readBytes(contentLength,
+            auto chunkResult = _socket->readBytes(static_cast<size_t>(contentLength),
                                                   args->onProgressCallback,
                                                   args->onChunkCallback,
                                                   isCancellationRequested);
@@ -507,7 +507,7 @@ namespace ix
 
             if (!args->onChunkCallback)
             {
-                payload.reserve(contentLength);
+                payload.reserve(static_cast<size_t>(contentLength));
                 payload += *chunkResult;
             }
         }
